+++ v2/api/storage/v1alpha1api20210401storage/storage_accounts_management_policy_types_gen_test.go
@@ -1536,7 +1536,7 @@ func DateAfterCreationStatusGenerator() gopter.Gen {
 
 // AddIndependentPropertyGeneratorsForDateAfterCreationStatus is a factory method for creating gopter generators
 func AddIndependentPropertyGeneratorsForDateAfterCreationStatus(gens map[string]gopter.Gen) {
-	gens["DaysAfterCreationGreaterThan"] = gen.PtrOf(gen.Int())
+	gens["DaysAfterCreationGreaterThan"] = gen.PtrOf(gen.Float64())
 }
 
 func Test_DateAfterModification_WhenSerializedToJson_DeserializesAsEqual(t *testing.T) {
